fix: cleanup workspace when switch setting panel
