+++ apps/web/src/adapters/local/index.tsx
@@ -113,7 +113,6 @@ export const LocalAdapter: WorkspaceAdapter<WorkspaceFlavour.LOCAL> = {
         <NewWorkspaceSettingDetail
           onDeleteWorkspace={onDeleteWorkspace}
           workspace={currentWorkspace}
-          key={currentWorkspace.id}
           onTransferWorkspace={onTransformWorkspace}
         />
       );
