+++ apps/web/src/providers/modal-provider.tsx
@@ -3,12 +3,14 @@ import {
   rootCurrentWorkspaceIdAtom,
   rootWorkspacesMetadataAtom,
 } from '@affine/workspace/atom';
+import { assertExists } from '@blocksuite/global/utils';
 import { arrayMove } from '@dnd-kit/sortable';
 import { useAtom, useSetAtom } from 'jotai';
 import { useRouter } from 'next/router';
-import type { ReactElement } from 'react';
+import type { FC, ReactElement } from 'react';
 import { lazy, Suspense, useCallback, useTransition } from 'react';
 
+import type { SettingAtom } from '../atoms';
 import {
   openCreateWorkspaceModalAtom,
   openDisableCloudAlertModalAtom,
@@ -16,9 +18,16 @@ import {
   openSettingModalAtom,
   openWorkspacesModalAtom,
 } from '../atoms';
+import { useCurrentWorkspace } from '../hooks/current/use-current-workspace';
 import { useRouterHelper } from '../hooks/use-router-helper';
 import { useWorkspaces } from '../hooks/use-workspaces';
 import type { AllWorkspace } from '../shared';
+
+const SettingModal = lazy(() =>
+  import('../components/affine/setting-modal').then(module => ({
+    default: module.SettingModal,
+  }))
+);
 
 const WorkspaceListModal = lazy(() =>
   import('../components/pure/workspace-list-modal').then(module => ({
@@ -46,7 +55,40 @@ const OnboardingModal = lazy(() =>
   }))
 );
 
+export const Setting: FC = () => {
+  const [currentWorkspace] = useCurrentWorkspace();
+  const [{ open, workspace, activeTab }, setOpenSettingModalAtom] =
+    useAtom(openSettingModalAtom);
+  assertExists(currentWorkspace);
+
+  const onSettingClick = useCallback(
+    ({
+      activeTab,
+      workspace,
+    }: Pick<SettingAtom, 'activeTab' | 'workspace'>) => {
+      setOpenSettingModalAtom(prev => ({ ...prev, activeTab, workspace }));
+    },
+    [setOpenSettingModalAtom]
+  );
+
+  return (
+    <SettingModal
+      open={open}
+      activeTab={activeTab}
+      workspace={workspace}
+      onSettingClick={onSettingClick}
+      setOpen={useCallback(
+        open => {
+          setOpenSettingModalAtom(prev => ({ ...prev, open }));
+        },
+        [setOpenSettingModalAtom]
+      )}
+    />
+  );
+};
+
 export function CurrentWorkspaceModals() {
+  const [currentWorkspace] = useCurrentWorkspace();
   const [openDisableCloudAlertModal, setOpenDisableCloudAlertModal] = useAtom(
     openDisableCloudAlertModalAtom
   );
@@ -75,6 +117,7 @@ export function CurrentWorkspaceModals() {
           />
         </Suspense>
       )}
+      {currentWorkspace && <Setting />}
     </>
   );
 }
